xsplice: Add support for shadow variables.
Shadow variables are a piece of infrastructure to be used by xsplice
modules. They are used to attach a new piece of data to an existing
structure in memory.

Martin Pohlack mentions that using small primers in the SHADOW_SLOTS
has the advantage of having a simple hash function, but cautions
that using round numbers (such as 256) will give "lot's of hash collisions
at the price of very fast hash computations as compilers, linkers, and
memory allocators tend to align starting addresses." (Martin)
Hence we pick a oddball size.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>
---
Cc: Ian Jackson <ian.jackson@eu.citrix.com>
Cc: Jan Beulich <jbeulich@suse.com>
Cc: Keir Fraser <keir@xen.org>
Cc: Tim Deegan <tim@xen.org>

v4: Add Copyright
v5: Change SHADOW_SLOTS to 257 per Martin's suggestion.
v6: Moving to end of series.
---