livepatch: Add support for shadow variables.
Shadow variables are a piece of infrastructure to be used by livepatch
modules. They are used to attach a new piece of data to an existing
structure in memory.

Martin Pohlack mentions that using small primers in the SHADOW_SLOTS
has the advantage of having a simple hash function, but cautions
that using round numbers (such as 256) will give "lot's of hash collisions
at the price of very fast hash computations as compilers, linkers, and
memory allocators tend to align starting addresses." (Martin)
Hence we pick a oddball size.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>
Signed-off-by: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>